requirements.txt
Django==1.11.14
django-map-widgets==0.1.9
psycopg2==2.7.5
pytz==2018.5
